# docs/swagger.yaml
basePath: /api
consumes:
- application/json
definitions:
  dto.EmailDTO:
    properties:
      email:
        example: myemail@gmail.com
        type: string
    required:
    - email
    type: object
  dto.EmailWithCodeDTO:
    properties:
      code:
        example: I1ELB
        type: string
      email:
        example: myemail@gmail.com
        type: string
      theme:
        default: SYSTEM
        example: DARK
        type: string
    required:
    - code
    - email
    type: object
  dto.EmailWithPasswordDTO:
    properties:
      email:
        example: myemail@gmail.com
        type: string
      password:
        example: onkr3451
        maxLength: 20
        minLength: 4
        type: string
      theme:
        default: SYSTEM
        example: DARK
        type: string
    required:
    - email
    - password
    type: object
  dto.MyUserDTO:
    description: User detail information
    properties:
      biography:
        example: I'd like to relax
        type: string
      email:
        example: myemail@gmail.com
        type: string
      firstName:
        example: Tele
        type: string
      friendsIds:
        example:
        - tldtb
        - kigfv
        items:
          type: string
        type: array
      isEmailVerified:
        type: boolean
      language:
        type: string
      lastName:
        example: phone
        type: string
      name:
        example: bobbas
        type: string
      role:
        type: string
      theme:
        type: string
    type: object
  dto.UpdateEmailDTO:
    properties:
      newEmail:
        example: myemail@gmail.com
        type: string
      password:
        example: onkr3451
        maxLength: 20
        minLength: 4
        type: string
    required:
    - newEmail
    - password
    type: object
  dto.UpdatePasswordDTO:
    properties:
      code:
        example: I1ELB
        type: string
      email:
        example: myemail@gmail.com
        type: string
      newPassword:
        example: onkr3451
        maxLength: 20
        minLength: 4
        type: string
    required:
    - code
    - email
    - newPassword
    type: object
  dto.UpdateUserDTO:
    properties:
      biography:
        example: I'd like to relax
        type: string
      firstName:
        example: Tele
        maxLength: 30
        type: string
      language:
        type: string
      lastName:
        example: phone
        maxLength: 30
        type: string
      theme:
        type: string
    type: object
  dto.UserDTO:
    description: User's main information
    properties:
      biography:
        example: I'd like to relax
        type: string
      firstName:
        example: Tele
        type: string
      friendsIds:
        example:
        - tldtb
        - kigfv
        items:
          type: string
        type: array
      lastName:
        example: phone
        type: string
      name:
        example: bobbas
        type: string
      role:
        type: string
    type: object
  exceptions.MyError:
    description: All native errors must be this type
    properties:
      advice:
        example: Try to send request later
        type: string
      message:
        example: Exception was occurred
        type: string
    type: object
host: localhost:3000
info:
  contact: {}
  description: It's an API interacting with You Together using Golang
  title: You Together API
  version: "1.0"
paths:
  /{name}:
    get:
      description: Returns type of object (NOT WORKING)
      parameters:
      - description: Name of something
        in: path
        name: name
        required: true
        type: string
      responses:
        "200":
          description: string type with object
          schema:
            $ref: '#/definitions/dto.UserDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exceptions.MyError'
        "404":
          description: User doesn't exist
          schema:
            $ref: '#/definitions/exceptions.MyError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exceptions.MyError'
      summary: Get type of the user (NOT WORKING)
      tags:
      - Get
  /auth/email/sign-in:
    post:
      description: Compare the secret code with the previously sent codes. If at least
        one matches, create session of the user. If the user does not exist, create
        a new user
      parameters:
      - description: User's email, secret code and device
        in: body
        name: user_info
        required: true
        schema:
          $ref: '#/definitions/dto.EmailWithCodeDTO'
      - description: User's language
        enum:
        - EN
        - RU
        in: header
        name: Accept-Language
        type: string
      responses:
        "200":
          description: user's session
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exceptions.MyError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exceptions.MyError'
      summary: Sign in by email
      tags:
      - Authorization
  /auth/password/sign-in:
    post:
      description: Compare the user's password with an existing user's password. If
        it matches, create session of the user. If the user does not exist, create
        a new user
      parameters:
      - description: User's email, password
        in: body
        name: EmailWithPasswordDTO
        required: true
        schema:
          $ref: '#/definitions/dto.EmailWithPasswordDTO'
      - description: User's language
        enum:
        - EN
        - RU
        in: header
        name: Accept-Language
        type: string
      responses:
        "200":
          description: user's session
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exceptions.MyError'
        "404":
          description: Password is not registered for this account
          schema:
            $ref: '#/definitions/exceptions.MyError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exceptions.MyError'
      summary: Sign in by password
      tags:
      - Authorization
  /email/send-code:
    post:
      description: Send a secret 5-digit code to the specified email
      parameters:
      - description: User's email
        in: body
        name: email
        required: true
        schema:
          $ref: '#/definitions/dto.EmailDTO'
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exceptions.MyError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exceptions.MyError'
      summary: Send code to the user's email
      tags:
      - Email
  /session:
    delete:
      description: Make the user's session invalid (can accept cookie)
      responses:
        "200":
          description: OK
      summary: Delete user's session
      tags:
      - Sessions
    get:
      description: Return detail information about the user (cookie required)
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MyUserDTO'
        "401":
          description: User isn't logged in
          schema:
            $ref: '#/definitions/exceptions.MyError'
        "404":
          description: User doesn't exist
          schema:
            $ref: '#/definitions/exceptions.MyError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exceptions.MyError'
      summary: Get detail information about user by session
      tags:
      - Sessions
  /user:
    patch:
      description: Change user's main information (cookie required)
      parameters:
      - description: New user Information
        in: body
        name: UpdateUserDTO
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserDTO'
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exceptions.MyError'
        "401":
          description: User isn't logged in
          schema:
            $ref: '#/definitions/exceptions.MyError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exceptions.MyError'
      summary: Update user's information
      tags:
      - User Update
  /user/{username}:
    get:
      description: Return main information about the user. If the user tries to find
        out information about himself, return detailed information about the user
        (can accept cookie)
      parameters:
      - description: Name of the user
        in: path
        name: username
        required: true
        type: string
      responses:
        "200":
          description: main info, without cookie or dto.MyUserDTO
          schema:
            $ref: '#/definitions/dto.UserDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exceptions.MyError'
        "404":
          description: User doesn't exist
          schema:
            $ref: '#/definitions/exceptions.MyError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exceptions.MyError'
      summary: Get main information about the user
      tags:
      - User Get
  /user/check-name/{username}:
    get:
      description: Status 200 if username not used or 403 if username already used
      parameters:
      - description: Name of the user
        in: path
        name: username
        required: true
        type: string
      responses:
        "200":
          description: name isn't used
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exceptions.MyError'
        "403":
          description: name already used
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exceptions.MyError'
      summary: Check username on exist
      tags:
      - User Get
  /user/email:
    patch:
      description: Change user's email by password (cookie required)
      parameters:
      - description: user's password and new email
        in: body
        name: UpdateEmailDTO
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateEmailDTO'
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exceptions.MyError'
        "401":
          description: User isn't logged in
          schema:
            $ref: '#/definitions/exceptions.MyError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exceptions.MyError'
      summary: Update user's email
      tags:
      - User Update
  /user/name:
    patch:
      description: Change user's username (cookie required)
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exceptions.MyError'
        "401":
          description: User isn't logged in
          schema:
            $ref: '#/definitions/exceptions.MyError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exceptions.MyError'
      summary: Update user's username
      tags:
      - User Update
  /user/password:
    patch:
      description: Change user's password by email (cookie required)
      parameters:
      - description: user's email, code and new password
        in: body
        name: UpdatePasswordDTO
        required: true
        schema:
          $ref: '#/definitions/dto.UpdatePasswordDTO'
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exceptions.MyError'
        "401":
          description: User isn't logged in
          schema:
            $ref: '#/definitions/exceptions.MyError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exceptions.MyError'
      summary: Update user's password
      tags:
      - User Update
produces:
- application/json
schemes:
- http
swagger: "2.0"